+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ## [Unreleased]
 ### Changed
 - `CodeGenerator`, `Validation`, templates: added optimisation of string length checks
+- `CodeGenerator`, `ItemConstraints`, templates: added checks on nested arrays
 
 ## [0.24.1] - 2021-02-03
 ### Changed