+++ CHANGELOG.md
@@ -2,11 +2,12 @@
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
-## [Unreleased]
+## [0.25] - 2021-02-28
 ### Changed
 - `CodeGenerator`, `Validation`, templates: added optimisation of string length checks
 - `CodeGenerator`, `ItemConstraints`, templates: added checks on nested arrays
 - `CodeGenerator`: added output filter to remove excess newlines
+- `CodeGenerator`, `Constraints`, `SystemClass`, templates: added support for `uniqueItems`
 
 ## [0.24.1] - 2021-02-03
 ### Changed